scripts/ml/requirements.txt
# requirements.txt for PSA Grade Prediction Project
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
requests>=2.25.0
tqdm>=4.62.0